+++ snapatac2-python/src/export.rs
@@ -1,5 +1,5 @@
 use crate::utils::AnnDataLike;
-use snapatac2_core::{export::{Exporter, Normalization}, utils};
+use snapatac2_core::{export::{Exporter, Normalization, CoverageOutputFormat}, utils};
 
 use std::ops::Deref;
 use anndata::Backend;
@@ -34,15 +34,15 @@ pub fn export_fragments(
 }
 
 #[pyfunction]
-pub fn export_bedgraph(
+pub fn export_coverage(
     anndata: AnnDataLike,
     group_by: Vec<&str>,
     resolution: usize,
     dir: PathBuf,
     prefix: &str,
     suffix:&str,
+    output_format: &str,
     selections: Option<HashSet<&str>>,
-    smooth_length: Option<u64>,
     blacklist: Option<PathBuf>,
     normalization: Option<&str>,
     ignore_for_norm: Option<HashSet<&str>>,
@@ -51,6 +51,7 @@ pub fn export_bedgraph(
     compression: Option<&str>,
     compression_level: Option<u32>,
     temp_dir: Option<PathBuf>,
+    num_threads: Option<usize>,
 ) -> Result<HashMap<String, PathBuf>> {
     let black: Option<BedTree<()>> = blacklist.map(|black| {
         Reader::new(utils::open_file_for_read(black), None)
@@ -60,33 +61,16 @@ pub fn export_bedgraph(
     });
 
     let normalization = normalization.map(|x| Normalization::from_str(x).unwrap());
+    let output_format = CoverageOutputFormat::from_str(output_format).unwrap();
 
     macro_rules! run {
         ($data:expr) => {
-            $data.export_bedgraph(
-                &group_by, selections, resolution, smooth_length, black.as_ref(), normalization,
+            $data.export_coverage(
+                &group_by, selections, resolution, black.as_ref(), normalization,
                 ignore_for_norm.as_ref(), min_frag_length, max_frag_length, dir, prefix,
-                suffix, compression.map(|x| utils::Compression::from_str(x).unwrap()), compression_level, temp_dir
+                suffix, output_format, compression.map(|x| utils::Compression::from_str(x).unwrap()),
+                compression_level, temp_dir, num_threads,
             )
-        }
-    }
-    crate::with_anndata!(&anndata, run)
-}
-
-
-#[pyfunction]
-pub fn export_bigwig(
-    anndata: AnnDataLike,
-    group_by: Vec<&str>,
-    selections: Option<HashSet<&str>>,
-    resolution: usize,
-    dir: PathBuf,
-    prefix: &str,
-    suffix: &str,
-) -> Result<HashMap<String, PathBuf>> {
-    macro_rules! run {
-        ($data:expr) => {
-            $data.export_bigwig(&group_by, selections, resolution, dir, prefix, suffix)
         }
     }
     crate::with_anndata!(&anndata, run)